Extract the Repository trait into a separate module
Break `repository.rs` into `repository/interface.rs` (trait + generic tests)
and `repository/postgres.rs`) (Postgres implementation) in preparation for
building a SQLite backend.
