+++ src/repository/mod.rs
@@ -1,0 +1,2 @@
+pub mod interface;
+pub mod postgres;
